[PVR] Refactor CPVRGUIActions step 10: Move remaining actions to own class.
